azure_functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity
openai
azure-keyvault-secrets
azure-storage-blob
azure-search-documents==11.4.0b11
azure-ai-formrecognizer==3.3.2
tiktoken==0.5.2
openai[datalib]==1.3.7
httpx==0.27.2
tenacity==8.2.3
azure-core
python-dotenv
pyodbc
azure-ai-textanalytics
applicationinsights
bs4
langchain
beautifulsoup4
html2text
sqlalchemy